better handle deleteProduct not found, keep PORT in .env

+++ backend/controller/product.controller.js
@@ -33,12 +33,17 @@ export const deleteProduct = async (req, res) => {
     const {id} = req.params; // get the id of the product to be deleted
     console.log("id:" + id); //show the id
 
+    // handle product not found
+    if (!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(404).json({ success: false, message: 'Product not found.' });
+    }
+
     try {
         await Product.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "product is deleted successfully"});
     } catch (error) {
         console.error("Error deleting product: " + error.message); // show on server log
-        res.status(404).json({ success: false, message: 'Product not found.' }); // show on client side
+        res.status(500).json({ success: false, message: 'Server Error.' }); // show on client side
     }
 };
